backend/pyproject.toml
[tool.poetry]
name        = "nosql-sim-backend"
version     = "0.2.0"
description = "Backend for NoSQL Simulator"
authors     = ["Yuri Adzhem"]
readme      = "../README.md"
package-mode = false

[tool.poetry.dependencies]
python   = "^3.12"

# Web framework
fastapi  = "^0.115.0"
uvicorn  = { extras = ["standard"], version = "^0.32.0" }
pydantic = { extras = ["email"], version = "^2.9.0" }
pydantic-settings = "^2.6.0"

# DB drivers / ORM
sqlalchemy = "^2.0.36"
asyncpg    = "^0.30.0"
alembic    = "^1.14.0"

# Auth
bcrypt        = "^4.2.0"
"python-jose" = { extras = ["cryptography"], version = "^3.3.0" }
"python-multipart" = "^0.0.12"

# NoSQL drivers (для песочниц)
motor    = "^3.6.0"
redis    = "^5.2.0"
neo4j    = "^5.27.0"
"cassandra-driver" = "^3.29.0"

# Utils
"python-dotenv" = "^1.0.1"
httpx           = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest         = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff           = "^0.8.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length    = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
